Migrate apply code to standalone module
